requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.50
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
anthropic>=0.40.0
pydantic>=2.8.0
python-multipart>=0.0.9
sse-starlette>=2.1.0
asyncpg>=0.29.0
